+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/graphql/types/GraphQLDayOfWeek.kt
@@ -1,0 +1,14 @@
+package com.leftindust.mockingbird.graphql.types
+
+import com.expediagroup.graphql.generator.annotations.GraphQLName
+
+@GraphQLName("DayOfWeek")
+enum class GraphQLDayOfWeek {
+    Mon,
+    Tue,
+    Wed,
+    Thu,
+    Fri,
+    Sat,
+    Sun;
+}
